# backend/swagger.yaml
openapi: 3.0.0
info:
  title: WalletLog API
  description: |
    Personal expense tracking API.
    Tracks income/expense **transactions**, groups them by **categories**, and
    enforces monthly **budgets** with overspend detection.

    ## Authentication
    Most endpoints require a Bearer access token (JWT, HS256, 15-min TTL).
    Click the **Authorize** button and paste the access token returned by
    `POST /auth/login`. The refresh token lives in an `httpOnly` cookie and
    is rotated automatically by the client.
  version: 1.1.0
  contact:
    name: Sertac Akalin
servers:
  - url: http://localhost:3000/api
    description: Local development

tags:
  - name: Auth
  - name: Categories
  - name: Transactions
  - name: Budgets

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Create a new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterInput' }
      responses:
        '201':
          description: Account created, refresh cookie set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }

  /auth/login:
    post:
      tags: [Auth]
      summary: Exchange credentials for an access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200':
          description: Logged in, refresh cookie set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate the refresh token, return a new access token
      security: []
      description: Reads the refresh token from the `refresh_token` httpOnly cookie.
      responses:
        '200':
          description: New access token + rotated refresh cookie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the refresh token and clear the cookie
      security: []
      responses:
        '204': { description: Logged out (no content) }

  /auth/me:
    get:
      tags: [Auth]
      summary: Get the current user
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /categories:
    get:
      tags: [Categories]
      summary: List the current user's categories
      responses:
        '200':
          description: Array of categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [Categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }

  /categories/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Categories]
      summary: Get a category by ID
      responses:
        '200':
          description: Category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Categories]
      summary: Update a category
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryInput' }
      responses:
        '200':
          description: Updated category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Categories]
      summary: Delete a category
      responses:
        '204': { description: Deleted (no content) }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  /transactions:
    get:
      tags: [Transactions]
      summary: List transactions (with optional filters)
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [income, expense] }
        - in: query
          name: category_id
          schema: { type: integer, minimum: 1 }
        - in: query
          name: startDate
          schema: { type: string, format: date }
          example: "2026-05-01"
        - in: query
          name: endDate
          schema: { type: string, format: date }
          example: "2026-05-31"
      responses:
        '200':
          description: Array of transactions (latest first)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /transactions/summary:
    get:
      tags: [Transactions]
      summary: Monthly income / expense / balance summary
      parameters:
        - in: query
          name: month
          required: true
          schema: { type: integer, minimum: 1, maximum: 12 }
        - in: query
          name: year
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Aggregated totals for the requested month
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthlySummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /transactions/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Transactions]
      summary: Get transaction by ID
      responses:
        '200':
          description: Transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Transactions]
      summary: Update a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionInput' }
      responses:
        '200':
          description: Updated transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Transactions]
      summary: Delete a transaction
      responses:
        '204': { description: Deleted (no content) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  /budgets:
    get:
      tags: [Budgets]
      summary: List the current user's budgets
      responses:
        '200':
          description: Array of budgets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Budget' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Budgets]
      summary: Create a new budget
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BudgetInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Budget' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }

  /budgets/status:
    get:
      tags: [Budgets]
      summary: Budget status (spent / remaining / exceeded) for a month
      parameters:
        - in: query
          name: month
          required: true
          schema: { type: integer, minimum: 1, maximum: 12 }
        - in: query
          name: year
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Per-category budget status
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BudgetStatus' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /budgets/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Budgets]
      summary: Get budget by ID
      responses:
        '200':
          description: Budget
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Budget' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Budgets]
      summary: Update a budget limit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [limit_amount]
              properties:
                limit_amount: { type: number, minimum: 0.01 }
      responses:
        '200':
          description: Updated budget
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Budget' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Budgets]
      summary: Delete a budget
      responses:
        '204': { description: Deleted (no content) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Not authenticated, or invalid/expired access token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: "Title is required" }

    User:
      type: object
      properties:
        id:           { type: integer, example: 1 }
        email:        { type: string,  example: "deniz@example.com" }
        display_name: { type: string,  nullable: true, example: "Deniz" }

    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:       { type: string, example: "deniz@example.com" }
        password:    { type: string, minLength: 8, maxLength: 72, example: "Sup3rSecret" }
        displayName: { type: string, maxLength: 80, example: "Deniz" }

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, example: "deniz@example.com" }
        password: { type: string, example: "Sup3rSecret" }

    AuthResponse:
      type: object
      properties:
        user:        { $ref: '#/components/schemas/User' }
        accessToken: { type: string, example: "eyJhbGciOi..." }

    Category:
      type: object
      properties:
        id:         { type: integer, example: 1 }
        user_id:    { type: integer, example: 1 }
        name:       { type: string,  example: "Food" }
        color:      { type: string,  example: "#3498db" }
        created_at: { type: string,  format: date-time }

    CategoryInput:
      type: object
      required: [name]
      properties:
        name:  { type: string,  maxLength: 100, example: "Food" }
        color: { type: string,  pattern: '^#[0-9A-Fa-f]{6}$', example: "#3498db" }

    Transaction:
      type: object
      properties:
        id:            { type: integer, example: 42 }
        user_id:       { type: integer, example: 1 }
        title:         { type: string,  example: "Grocery shopping" }
        amount:        { type: number,  example: 125.50 }
        type:          { type: string,  enum: [income, expense] }
        category_id:   { type: integer, nullable: true }
        category_name: { type: string,  nullable: true }
        date:          { type: string,  format: date }
        note:          { type: string,  nullable: true }
        created_at:    { type: string,  format: date-time }

    TransactionInput:
      type: object
      required: [title, amount, type, date]
      properties:
        title:       { type: string,  maxLength: 200 }
        amount:      { type: number,  minimum: 0.01 }
        type:        { type: string,  enum: [income, expense] }
        category_id: { type: integer, nullable: true }
        date:        { type: string,  format: date, example: "2026-05-15" }
        note:        { type: string,  nullable: true }

    MonthlySummary:
      type: object
      properties:
        income:  { type: number, example: 5000 }
        expense: { type: number, example: 1450 }
        balance: { type: number, example: 3550 }

    Budget:
      type: object
      properties:
        id:            { type: integer }
        user_id:       { type: integer }
        category_id:   { type: integer }
        category_name: { type: string }
        month:         { type: integer, minimum: 1, maximum: 12 }
        year:          { type: integer }
        limit_amount:  { type: number }

    BudgetInput:
      type: object
      required: [category_id, month, year, limit_amount]
      properties:
        category_id:  { type: integer, minimum: 1 }
        month:        { type: integer, minimum: 1, maximum: 12 }
        year:         { type: integer }
        limit_amount: { type: number, minimum: 0.01 }

    BudgetStatus:
      type: object
      properties:
        budget_id:   { type: integer }
        category_id: { type: integer }
        category:    { type: string,  example: "Food" }
        limit:       { type: number,  example: 500 }
        spent:       { type: number,  example: 320 }
        remaining:   { type: number,  example: 180 }
        exceeded:    { type: boolean, example: false }